@@ -1,73 +1,56 @@
 module Client
   class CallersController < ClientController
     include TimeZoneHelper
-    before_filter :full_access, :except =>[:reassign_to_campaign]
-    include DeletableController
-    include ApplicationHelper::TimeUtils
     skip_before_filter :check_login, :only => [:reassign_to_campaign]
-    before_filter :load_campaigns, :except => [:index,:destroy,:reassign_to_campaign]
+    before_filter :full_access, :except => [:reassign_to_campaign]
+    before_filter :load_and_verify_caller, :except => [:index, :new, :create, :reassign_to_campaign, :usage, :call_details, :type_name, :deleted]
+    before_filter :load_campaigns, :except => [:index, :destroy, :reassign_to_campaign, :usage, :call_details, :type_name, :deleted]
 
-    def type_name
-      'caller'
-    end
+    respond_to :html, :json
 
     def index
-      @breadcrumb="Callers"
-      @callers = Caller.where(:active => true, :account_id => account.id).order(:name).paginate(:page => params[:page])
+      @callers = account.callers.active.paginate(:page => params[:page])
+      respond_with @callers
     end
 
     def new
-      @breadcrumb=[{"Callers"=>"/client/callers"}, "Add Caller"]
-      @caller = Caller.new(:is_phones_only => params[:is_phones_only])
+      @caller = account.callers.new(:is_phones_only => params[:is_phones_only])
+      load_caller_groups
+      respond_with @caller
     end
 
     def show
-      @breadcrumb=[{"Callers"=>"/client/callers"}, "Add Caller"]
-      @caller = account.callers.find_by_id(params[:id])
-      @caller ||= Caller.new
-      if @caller.new_record?
-        flash[:error] = ["No caller found. Create a new caller?"]
-        render :action => "new"
+      respond_with @caller do |format|
+        format.html {redirect_to edit_client_caller_path(@caller)}
       end
     end
 
+    def edit
+      load_caller_groups
+      respond_with @caller
+    end
 
     def update
-      @caller = account.callers.find_by_id(params[:id])
-      if @caller.is_on_call? && (params[:caller][:campaign_id] != @caller.campaign.id)
-        flash_message(:error, "This caller is logged in and so can't be changed to a new campaign from this screen. To reassign them to a new campaign, please use the Monitor tab.")
-        render :action=>"new"
-      else
-        if @caller.update_attributes(params[:caller])
-          flash_message(:notice, "Caller updated")
-          redirect_to :action=>"index"
-        else
-          render :action=>"new"
-        end
+      save_caller
+      respond_with @caller, location: client_callers_path do |format|
+        format.json {render :json => {message: 'Caller updated'}, status: :ok} if @caller.errors.empty?
       end
     end
 
     def create
-      @caller = Caller.new(params[:caller])
-      @caller.account_id = account.id
-      if @caller.save
-        flash_message(:notice, "Caller saved")
-        redirect_to :action=>"index"
-      else
-        render :action=>"new"
-      end
+      @caller = account.callers.new
+      save_caller
+      respond_with @caller, location: client_callers_path
     end
 
     def destroy
-      @caller = account.callers.find_by_id(params[:id])
-      if !@caller.blank?
-        @caller.active = false
-        @caller.save
+      @caller.active = false
+      @caller.save ? flash_message(:notice, "Caller deleted") : flash_message(:error, @caller.errors.full_messages.join)
+      respond_with @caller, location: client_callers_path do |format|
+        format.json {render :json => {message: 'Caller deleted'}, :status => :ok} if @caller.errors.empty?
       end
-      flash_message(:notice, "Caller deleted")
-      redirect_to :action=>"index"
     end
-    
+
     def reassign_to_campaign
       caller = Caller.find_by_id(params[:id])
       caller.update_attributes(:campaign_id => params[:campaign_id])
@@ -88,14 +71,50 @@ module Client
       @caller = Caller.find(params[:id])
       @campaigns = account.campaigns.manual.for_caller(@caller)
       @campaign = @campaigns.find_by_id(params[:campaign_id]) || @caller.caller_sessions.last.try(:campaign) || @caller.campaign
-      @from_date, @to_date = set_date_range_callers(@campaign, @caller, params[:from_date], params[:to_date])   
+      @from_date, @to_date = set_date_range_callers(@campaign, @caller, params[:from_date], params[:to_date])
       @answered_call_stats = @caller.answered_call_stats(@from_date, @to_date, @campaign)
       @questions_and_responses = @campaign.try(:questions_and_responses) || {}
     end
+    
+    def deleted
+      @callers = Caller.deleted.for_account(account).paginate(:page => params[:page], :order => 'id desc')
+      respond_with @callers do |format|
+        format.html{render 'client/callers/deleted'}
+        format.json {render :json => @callers.to_json}
+      end      
+    end
+
+    def type_name
+      'caller'
+    end
 
     private
+
+    def load_and_verify_caller
+      begin
+        @caller = Caller.find(params[:id])
+      rescue ActiveRecord::RecordNotFound => e
+        render :json=> {"message"=>"Resource not found"}, :status => :not_found
+        return
+      end
+      if @caller.account != account
+        render :json => {message: 'Cannot access caller'}, :status => :unauthorized
+        return
+      end
+    end
+
     def load_campaigns
       @campaigns = account.campaigns.manual.active
-    end        
+    end
+
+    def load_caller_groups
+      @caller_groups = account.caller_groups
+    end
+
+    def save_caller
+      load_campaigns
+      load_caller_groups
+      flash_message(:notice, "Caller saved") if @caller.update_attributes(params[:caller])
+    end
   end
 end
