@@ -1,6 +1,27 @@
 # Methods added to this helper will be available to all templates in the application.
 module ApplicationHelper
   include WhiteLabeling
+  
+  def flash_message(where, error_message)
+    if flash[where] and flash[where].class == Array
+      flash[where] = flash[where].concat [error_message]  # should not use <<. rails flash does not 'keep' them.
+    elsif flash[where] and flash[where].class == String
+      flash[where] = [flash[where], error_message]
+    else
+      flash[where] = [error_message]
+    end
+  end
+
+  def flash_now(where, error_message)
+    if flash.now[where] and flash.now[where].class == Array
+      flash.now[where] = flash.now[where].concat [error_message]  # should not use <<. rails flash does not 'keep' them.
+    elsif flash.now[where] and flash.now[where].class == String
+      flash.now[where] = [flash.now[where], error_message]
+    else
+      flash.now[where] = [error_message]
+    end
+  end
+  
 
   def cms(key)
     s = Seo.find_by_crmkey_and_active_and_version(key, 1, session[:seo_version])
@@ -58,7 +79,7 @@ module ApplicationHelper
 
 
   def client_controller?(controllerName)
-    ['client/accounts', 'client', 'voter_lists', 'monitor', 'client/campaigns', 'client/scripts', 'client/callers', 'client/reports', 'campaigns', 'scripts', 'broadcast', 'reports', 'home', 'blocked_numbers', 'monitors', 'messages'].include?(controllerName)
+    ['client/accounts', 'client', 'voter_lists', 'monitor', 'client/campaigns', 'client/scripts', 'client/callers', 'client/reports', 'campaigns', 'scripts', 'client/caller_groups', 'broadcast', 'reports', 'home', 'blocked_numbers', 'monitors', 'messages'].include?(controllerName)
   end
 
   ['title', 'full_title', 'phone', 'email', 'billing_link'].each do |value|
@@ -95,11 +116,11 @@ module ApplicationHelper
   def pusher_js
     javascript_include_tag("https://d3dy5gmtp8yhk7.cloudfront.net/1.11/pusher.min.js")
   end
-  
+
   def sanitize_dials(dial_count)
     dial_count.nil? ? 0 : dial_count
   end
-  
+
   def percent_dials(dial_count, total_dials)
     begin
       ((sanitize_dials(dial_count).to_f/total_dials)*100).round
@@ -107,7 +128,7 @@ module ApplicationHelper
       0
     end
   end
-  
+
 
 
   module TimeUtils
