@@ -4,27 +4,30 @@ class Caller < ActiveRecord::Base
   validates_format_of :email, :allow_blank => true, :with => /^([^@\s]+)@((?:[-a-z0-9]+\.)+[a-z]{2,})$/i, :message => "Invalid email"
   belongs_to :campaign
   belongs_to :account
+  belongs_to :caller_group
   has_many :caller_sessions
   has_many :caller_identities
   has_many :call_attempts
   has_many :answers
   before_create :create_uniq_pin
+  before_validation :assign_to_caller_group_campaign
   validates_uniqueness_of :email, :allow_nil => true
-  validates_presence_of :campaign_id
+  validates :campaign_id, presence: true
 
   scope :active, where(:active => true)
-  
+
   delegate :subscription_allows_caller?, :to => :account
   delegate :activated?, :to => :account
   delegate :funds_available?, :to => :account
+  delegate :as_time_zone, :to=> :campaign
 
   cattr_reader :per_page
   @@per_page = 25
-  
+
   def identity_name
     is_phones_only?  ? name : email
   end
-  
+
   def create_uniq_pin
     uniq_pin=0
     while uniq_pin==0 do
@@ -34,10 +37,10 @@ class Caller < ActiveRecord::Base
     end
     self.pin = uniq_pin
   end
-  
-  
 
-  
+
+
+
   def is_on_call?
     !caller_sessions.blank? && caller_sessions.on_call.size > 0
   end
@@ -68,7 +71,7 @@ class Caller < ActiveRecord::Base
     Twilio::Verb.new { |v| v.play("#{APP_URL}/wav/hold.mp3"); v.redirect(hold_call_path(:host => Settings.host, :port => Settings.port), :method => "GET")}.response
   end
 
-  def callin(campaign)    
+  def callin(campaign)
     response = TwilioClient.instance.account.calls.create(
         :from =>APP_NUMBER,
         :to => Settings.phone,
@@ -85,25 +88,25 @@ class Caller < ActiveRecord::Base
     return email unless email.blank?
     ''
   end
-  
+
   def info
     attributes.reject { |k, v| (k == "created_at") ||(k == "updated_at") }
   end
-  
-  
+
+
   def answered_call_stats(from, to, campaign)
     result = Hash.new
     question_ids = Answer.all(:select=>"distinct question_id", :conditions=>"campaign_id = #{campaign.id}")
     answer_count = Answer.select("possible_response_id").where("campaign_id = ? and caller_id = ?", campaign.id, self.id).within(from, to).group("possible_response_id").count
     total_answers = Answer.select("question_id").from("answers use index (index_answers_count_question_id)").where("campaign_id = ? and caller_id = ?",campaign.id, self.id).within(from, to).group("question_id").count
     questions = Question.where("id in (?)",question_ids.collect{|q| q.question_id})
-    questions.each do |question|        
+    questions.each do |question|
       result[question.text] = question.possible_responses.collect { |possible_response| possible_response.stats(answer_count, total_answers) }
       result[question.text] << {answer: "[No response]", number: 0, percentage:  0} unless question.possible_responses.find_by_value("[No response]").present?
     end
     result
   end
-  
+
   # def reassign_to_another_campaign(caller_session)
   #   return unless caller_session.attempt_in_progress.nil?
   #   if self.is_phones_only?
@@ -114,27 +117,33 @@ class Caller < ActiveRecord::Base
   #      else
   #        caller_session.reassign_caller_session_to_campaign
   #        if campaign.predictive_type == Campaign::Type::PREVIEW || campaign.predictive_type == Campaign::Type::PROGRESSIVE
-  #          caller_session.publish('conference_started', {}) 
+  #          caller_session.publish('conference_started', {})
   #        else
   #          caller_session.publish('caller_connected_dialer', {})
   #        end
   #      end
   # end
-  
-  
+
+
   def create_caller_session(session_key, sid, caller_type)
     if is_phones_only?
-      caller_session = PhonesOnlyCallerSession.create(on_call: false, available_for_call: false, session_key: session_key, campaign: campaign , sid: sid, starttime: Time.now, caller_type: caller_type)
+      caller_session = PhonesOnlyCallerSession.create(on_call: false, available_for_call: false, session_key: session_key, campaign: campaign , sid: sid, starttime: Time.now, caller_type: caller_type, state: 'initial')
     else
-      caller_session =  WebuiCallerSession.create(on_call: false, available_for_call: false, session_key: session_key, campaign: campaign , sid: sid, starttime: Time.now, caller_type: caller_type)
+      caller_session =  WebuiCallerSession.create(on_call: false, available_for_call: false, session_key: session_key, campaign: campaign , sid: sid, starttime: Time.now, caller_type: caller_type, state: 'initial')
     end
     caller_sessions << caller_session
     caller_session
   end
-  
-  def create_caller_identity(session_key)
-    caller_identities.create(session_key: session_key, pin: CallerIdentity.create_uniq_pin)    
-  end
-  
 
+  def create_caller_identity(session_key)
+    caller_identities.create(session_key: session_key, pin: CallerIdentity.create_uniq_pin)
+  end
+
+  private
+
+  def assign_to_caller_group_campaign
+    if caller_group_id_changed? && !caller_group_id.nil?
+      self.campaign_id = CallerGroup.find(caller_group_id).campaign_id
+    end
+  end
 end
