@@ -12,7 +12,7 @@ module LeadEvents
     def publish_voter_connected      
       unless caller_session.nil?
         EM.synchrony {
-          unless caller_session.caller.is_phones_only?
+          if !caller_session.caller.is_phones_only? && !voter.caller_session.nil?
             event_hash = campaign.voter_connected_event(self.call)        
             caller_deferrable = Pusher[caller_session.session_key].trigger_async(event_hash[:event], event_hash[:data].merge!(:dialer => campaign.type))
             caller_deferrable.callback {}
