@@ -25,7 +25,7 @@ class ReportDownloadJob
    
    def self.after_enqueue_scale_up(*args)
       workers_to_scale = HerokuResqueAutoScale::Scaler.working_job_count('report_download_worker_job') + HerokuResqueAutoScale::Scaler.pending_job_count('report_download_worker_job') - HerokuResqueAutoScale::Scaler.worker_count('report_download_worker_job')
-      if workers_to_scale > 0 && HerokuResqueAutoScale::Scaler.working_job_count('report_download_worker_job') < 3
+      if workers_to_scale > 0 && HerokuResqueAutoScale::Scaler.working_job_count('report_download_worker_job') < 6
         HerokuResqueAutoScale::Scaler.workers('report_download_worker_job', (HerokuResqueAutoScale::Scaler.working_job_count('report_download_worker_job') + HerokuResqueAutoScale::Scaler.pending_job_count('report_download_worker_job')))
       end
     end
