@@ -3,8 +3,9 @@ class VoterListUploadJob
   extend ::HerokuResqueAutoScale
   @queue = :voter_list_upload_worker_job
 
-   def self.perform(separator, column_headers, csv_to_system_map, filename, voter_list_name, campaign_id, account_id, domain, email, callback_url, strategy="webui")
-     job = VoterListJob.new(separator, column_headers, csv_to_system_map, filename, voter_list_name, campaign_id, account_id, domain, email, callback_url, strategy="webui")
+   def self.perform(voter_list_id, email, domain, callback_url, strategy="webui")
+     voter_list = VoterList.find(voter_list_id)
+     job = VoterListJob.new( voter_list.id , domain, email, callback_url, strategy="webui")
      job.perform
    end
    
@@ -14,7 +15,7 @@ class VoterListUploadJob
    
    def self.after_enqueue_scale_up(*args)
       workers_to_scale = HerokuResqueAutoScale::Scaler.working_job_count('voter_list_upload_worker_job') + HerokuResqueAutoScale::Scaler.pending_job_count('voter_list_upload_worker_job') - HerokuResqueAutoScale::Scaler.worker_count('voter_list_upload_worker_job')
-      if workers_to_scale > 0 && HerokuResqueAutoScale::Scaler.working_job_count('voter_list_upload_worker_job') < 3
+      if workers_to_scale > 0 && HerokuResqueAutoScale::Scaler.working_job_count('voter_list_upload_worker_job') < 6
         HerokuResqueAutoScale::Scaler.workers('voter_list_upload_worker_job', (HerokuResqueAutoScale::Scaler.working_job_count('voter_list_upload_worker_job') + HerokuResqueAutoScale::Scaler.pending_job_count('voter_list_upload_worker_job')))
       end
     end
