@@ -1,17 +1,17 @@
 var Scripts = function(){
-	
-  var self = this;		
+
+  var self = this;
 
   self.display_question_numbers();
   self.display_text_field_numbers();
-  self.display_script_text_numbers();          
+  self.display_script_text_numbers();
 
 
-  $("#script_elements").sortable({ cursor: 'crosshair' , axis: 'y' ,stop: function(event, ui) { 
+  $("#script_elements").sortable({ cursor: 'crosshair' , axis: 'y' ,stop: function(event, ui) {
     self.display_question_numbers();
-	self.display_script_text_numbers();
-	self.display_text_field_numbers();
-  } 
+  self.display_script_text_numbers();
+  self.display_text_field_numbers();
+  }
   });
 
 
@@ -24,31 +24,31 @@ var Scripts = function(){
     return true;
   });
 
-  $('#call_results').bind('insertion-callback',            
+  $('#call_results').bind('insertion-callback',
     function() {
       self.display_script_text_numbers();
       self.display_question_numbers();
-      self.display_text_field_numbers();                              
+      self.display_text_field_numbers();
       self.add_new_response_when_question_added();
   });
 
 
 
-  $('#call_results').bind('after-removal-callback',            
-    function() { 
-	  self.display_script_text_numbers();
+  $('#call_results').bind('after-removal-callback',
+    function() {
+    self.display_script_text_numbers();
       self.display_question_numbers();
-      self.display_text_field_numbers();                              
+      self.display_text_field_numbers();
       self.add_new_response_when_question_added();
-    
-	});
-  
+
+  });
+
 
 
 }
 
 Scripts.prototype.set_elements_order = function(){
-  var count = 1;    
+  var count = 1;
   $.each($('.script_element'), function(){
       $(this).val(count++);
   });
@@ -56,9 +56,9 @@ Scripts.prototype.set_elements_order = function(){
 
 
 Scripts.prototype.display_question_numbers = function(){
-  var question_count = 1;    
+  var question_count = 1;
   $.each($('.question_label'), function(){
-	if ($(this).parent('fieldset').attr('deleted') != "true") {
+  if ($(this).parent('fieldset').attr('deleted') != "true") {
       $(this).text("Question "+question_count++);
     }
   });
@@ -66,9 +66,9 @@ Scripts.prototype.display_question_numbers = function(){
 }
 
 Scripts.prototype.display_script_text_numbers = function(){
-  var script_text_count = 1;    
+  var script_text_count = 1;
   $.each($('.script_label'), function(){
-	if ($(this).parent('fieldset').attr('deleted') != "true") {
+  if ($(this).parent('fieldset').attr('deleted') != "true") {
       $(this).text("Script Text "+script_text_count++);
     }
   });
@@ -77,7 +77,7 @@ Scripts.prototype.display_script_text_numbers = function(){
 
 
 Scripts.prototype.set_question_order = function(){
-  var question_order = 1;    
+  var question_order = 1;
   $.each($('.question_order'), function(){
       $(this).val(question_order++);
   });
@@ -87,11 +87,11 @@ Scripts.prototype.set_question_order = function(){
 Scripts.prototype.set_possible_response_order = function(){
   responses = $('.possible_response_sortable')
   $.each(responses, function(){
-	var possible_response_order = 1;    
-	var possible_responses = $(this).find('.possible_response_order');
-	$.each($(possible_responses), function(){
-	  $(this).val(possible_response_order++);			
-	});	
+  var possible_response_order = 1;
+  var possible_responses = $(this).find('.possible_response_order');
+  $.each($(possible_responses), function(){
+    $(this).val(possible_response_order++);
+  });
   });
 
 }
@@ -99,12 +99,12 @@ Scripts.prototype.set_possible_response_order = function(){
 
 
 Scripts.prototype.mark_answered = function(){
-	question_ids = new Array();
-	$('#script_questions').find('.identity').each(function(){ 
-	   question_ids.push($(this).val());
-	})
-	questions_answered();
-	possible_response_answered(question_ids);
+  question_ids = new Array();
+  $('#script_questions').find('.identity').each(function(){
+     question_ids.push($(this).val());
+  })
+  questions_answered();
+  possible_response_answered(question_ids);
 
 }
 
@@ -112,8 +112,8 @@ Scripts.prototype.display_text_field_numbers = function(){
   var text_field_count = 1;
   $.each($('.text_field_label'), function(){
     if ($(this).parent('fieldset').attr('deleted') != "true") {
-      $(this).text("Text field "+text_field_count++);
-	}
+      $(this).text("Note "+text_field_count++);
+  }
   });
 }
 
@@ -121,12 +121,12 @@ Scripts.prototype.add_new_response_when_question_added = function(){
   $.each($('.question'), function(){
     if ($(this).find('.possible_response').length == 0) {
       $(this).children('.add_response').trigger('click')
-    }      
-  });    
+    }
+  });
 }
 
 Scripts.prototype.reorder_script_elements = function(){
-	
+
 }
 
 function question_delete(question_node){
@@ -135,53 +135,54 @@ if(($('#script_questions').children('.nested-fields').length - $('fieldset.quest
       return false;
     }
  else {
-	if( $($(question_node).parent('.question').children('.identity')[0]).attr('answered') == "true"){
+  if( $($(question_node).parent('.question').children('.identity')[0]).attr('answered') == "true"){
       var confirm_delete = confirm("This question has already been answered. Are you sure you want to delete it? You will lose all of its data.");
- 	  return confirm_delete
-	}
+     return confirm_delete
+  }
 }
-  return true;   
+  return true;
 }
 
 function possible_response_answered(question_ids){
   $.ajax({
-    url : "/client/scripts/possible_responses_answered",
-    data : {question_ids : question_ids },
+    url : "/client/scripts//possible_responses_answered",
+    data : {question_ids : question_ids, id: $("#script_id").val() },
     type : "GET",
-	async : false,
+  async : false,
     success : function(response) {
-		for (myKey in response["data"]){
-			$('#call_results').find('.possible_response_identity').each(function(){
-				if ($(this).val() == myKey) {					
-					$(this).attr('answered', true)
-				}
-			});
-		}
-	}
+    for (myKey in response["data"]){
+      $('#call_results').find('.possible_response_identity').each(function(){
+        if ($(this).val() == myKey) {
+          $(this).attr('answered', true)
+        }
+      });
+    }
+  }
   });
 }
 
 
 function questions_answered(){
+	var script_id = $('#script_id').val();
   $.ajax({
     url : "/client/scripts/questions_answered",
     data : {id : $("#script_id").val() },
     type : "GET",
-	async : false,
+  async : false,
     success : function(response) {
-		for (myKey in response["data"]){
-			$('#script_questions').find('.identity').each(function(){
-				if ($(this).val() == myKey) {					
-					$(this).attr('answered', true)
-				}
-			});
-		}
-	}
+    for (myKey in response["data"]){
+      $('#script_questions').find('.identity').each(function(){
+        if ($(this).val() == myKey) {
+          $(this).attr('answered', true)
+        }
+      });
+    }
+  }
   });
 }
 
 function possible_response_delete(response_node){
-	
+
   question = $(response_node).parents('.question')[0];
   if(( $(question).find('table.possible_response').length  - $(question).find('table.possible_response[deleted="true"]').length ) == 1) {
     alert("You must have at least one response.");
@@ -189,10 +190,10 @@ function possible_response_delete(response_node){
   }
   else{
     if( $($($(response_node).parents('.possible_response')[0]).children('.possible_response_identity')[0]).attr('answered') == "true"){
-	  var confirm_delete = confirm("This response has already been chosen. Are you sure you want to delete it? You will lose all of its data.");
-	  return confirm_delete
-	}
-  }    	
+    var confirm_delete = confirm("This response has already been chosen. Are you sure you want to delete it? You will lose all of its data.");
+    return confirm_delete
+  }
+  }
   return true;
 }
 
